style: Profile card updated with new and preferred design

+++ src/components/Pages/UserProfile.jsx
@@ -2,24 +2,27 @@ import React, { useState, useEffect } from 'react'
 import { Link, useParams } from 'react-router-dom'
 // import GithubContext from '../../Context/GithubContext'
 import { useGithubContext } from '../../hooks/useGithubContext'
-
-import { MdArrowBackIos } from 'react-icons/md'
-import { BiMap, BiGlobe, BiNote } from 'react-icons/bi'
-import {
-  FaCodepen,
-  FaStore,
-  FaUserFriends,
-  FaUsers,
-  FaGithub,
-  FaTwitter,
-} from 'react-icons/fa'
 import { getUserData } from '../../utils/axios'
-
 import Repositories from '../Layouts/Repositories'
+
+// icons
+import { MdArrowBackIos } from 'react-icons/md'
+import { BsGithub, BsTwitter } from 'react-icons/bs'
+
+import {
+  GoPeople,
+  GoOrganization,
+  GoLocation,
+  GoLink,
+  GoRepo,
+  GoLog,
+  GoPersonAdd,
+  GoCodeSquare,
+} from 'react-icons/go'
 
 export default function UserProfile() {
   const [getGithubUser, setGithubUser] = useState({})
-  const { setStore } = useGithubContext()
+  const { store, setStore } = useGithubContext()
 
   const { username } = useParams()
 
@@ -31,18 +34,28 @@ export default function UserProfile() {
       .catch((err) => console.log(err))
   }, [username])
 
+  if (store.loading) {
+    return <h1>Loading...</h1>
+  }
+
+  console.log(getGithubUser)
+
   return (
     <React.Fragment>
       {/* back button */}
-      <section className="my-3">
+      <section className="mt-3">
         {getGithubUser.type === 'User' ? (
           <Link
             to="/users"
             type="button"
-            className="btn  mt-4"
+            className="btn fw-medium"
             onClick={() => {
               window.sessionStorage.clear()
-              setStore('')
+              setStore({
+                username: '',
+                error: false,
+                loading: true,
+              })
             }}
           >
             <MdArrowBackIos size={20} /> Back to Search
@@ -51,10 +64,14 @@ export default function UserProfile() {
           <Link
             to="/orgs"
             type="button"
-            className="btn  mt-4"
+            className="btn fw-medium"
             onClick={() => {
               window.sessionStorage.clear()
-              setStore('')
+              setStore({
+                username: '',
+                error: false,
+                loading: true,
+              })
             }}
           >
             <MdArrowBackIos size={20} /> Back to Search
@@ -63,41 +80,65 @@ export default function UserProfile() {
       </section>
 
       {/* profile section */}
-      <section className="my-3">
+      <section className="mt-4">
         <article className="row mb-3">
-          <div className="col-md-4 text-center my-2">
+          <div className="col-md-4 text-center my-3">
             <img
               src={getGithubUser?.avatar_url}
               alt={getGithubUser?.login}
-              className="profile-card-img img-fluid overflow-hidden rounded-4 shadow-lg"
+              className="img-fluid overflow-hidden rounded-2 shadow-lg"
+              width={'65%'}
             />
           </div>
           <div className="col-md-8 my-2">
-            <div className="d-flex align-items-center">
-              <h2 className="me-3">{getGithubUser?.name}</h2>
+            <div className="d-flex gap-3 align-items-center">
+              <h1 className="h2 fw-bold mb-0">
+                {getGithubUser?.name ? getGithubUser?.name : 'NA'}
+              </h1>
 
-              {getGithubUser?.hireable ? (
-                <span className="badge bg-success text-light">Hireable</span>
-              ) : getGithubUser?.type === 'User' ? (
-                <span className="badge bg-danger text-light">Not-Hireable</span>
+              {getGithubUser?.type === 'User' ? (
+                getGithubUser?.hireable ? (
+                  <span className="badge bg-success text-light">Hireable</span>
+                ) : (
+                  <span className="badge bg-danger text-light">
+                    Not-Hireable
+                  </span>
+                )
               ) : null}
+
+              <a
+                type="button"
+                href={getGithubUser.html_url}
+                target="_blank"
+                rel="noopener noreferrer"
+                className="btn profile-github-button fw-medium"
+                title="visit github"
+              >
+                <BsGithub size={20} />
+              </a>
             </div>
-            <h5 className="mb-3">@ {getGithubUser?.login}</h5>
+            <p className="fw-medium">@{getGithubUser?.login}</p>
 
             {getGithubUser.bio && (
-              <h5 className="mb-3">
-                <BiNote /> {getGithubUser?.bio}
-              </h5>
+              <p className="fw-medium">
+                <GoLog /> {getGithubUser?.bio}
+              </p>
             )}
 
-            {getGithubUser.location && (
-              <h5 className="mb-3">
-                <BiMap /> {getGithubUser.location}
-              </h5>
+            {getGithubUser.company && (
+              <p className="fw-medium">
+                <GoOrganization /> {getGithubUser.company}
+              </p>
             )}
+            {getGithubUser.location && (
+              <p className="fw-medium">
+                <GoLocation /> {getGithubUser.location}
+              </p>
+            )}
+
             {getGithubUser.blog && (
-              <h5 className="mb-3">
-                <BiGlobe />{' '}
+              <p className="fw-medium">
+                <GoLink />{' '}
                 <a
                   href={getGithubUser.blog}
                   target="_blank"
@@ -105,63 +146,53 @@ export default function UserProfile() {
                 >
                   {getGithubUser.blog}
                 </a>
-              </h5>
+              </p>
             )}
+
             {getGithubUser.twitter_username && (
-              <h5 className="mb-3">
-                <FaTwitter />{' '}
+              <p className="fw-medium">
+                <BsTwitter />{' '}
                 <a
                   href={`https://twitter.com/${getGithubUser.twitter_username}`}
                   target="_blank"
                   rel="noopener noreferrer"
+                  className="link-card-highlight"
                 >
                   {getGithubUser.twitter_username}
                 </a>
-              </h5>
+              </p>
             )}
-
-            <div className="my-2">
-              <a
-                type="button"
-                href={getGithubUser.html_url}
-                target="_blank"
-                rel="noopener noreferrer"
-                className="btn profile-github-button"
-              >
-                <FaGithub size={22} /> Visit GitHub Profile
-              </a>
-            </div>
           </div>
         </article>
 
-        <article className="row g-1 py-3 row-cols-1 row-cols-md-2 row-cols-lg-4">
+        <article className="row g-1 py-3 row-cols-2 row-cols-lg-4">
           <div className="col d-flex align-items-center">
             <div className="icon-square text-bg-dark d-inline-flex align-items-center justify-content-center p-2 rounded me-3">
-              <FaUsers size={30} />
+              <GoPeople size={30} />
             </div>
             <div>
               <p className="m-0 p-0">Followers</p>
-              <p className="m-0 p-0 fw-bold fs-2">{getGithubUser?.followers}</p>
+              <p className="m-0 p-0 fw-bold fs-3">{getGithubUser?.followers}</p>
             </div>
           </div>
 
           <div className="col d-flex align-items-center">
             <div className="icon-square text-bg-dark d-inline-flex align-items-center justify-content-center p-2 rounded me-3">
-              <FaUserFriends size={30} />
+              <GoPersonAdd size={30} />
             </div>
             <div>
               <p className="m-0 p-0">Following</p>
-              <p className="m-0 p-0 fw-bold fs-2">{getGithubUser?.following}</p>
+              <p className="m-0 p-0 fw-bold fs-3">{getGithubUser?.following}</p>
             </div>
           </div>
 
           <div className="col d-flex align-items-center">
             <div className="icon-square text-bg-dark d-inline-flex align-items-center justify-content-center p-2 rounded me-3">
-              <FaCodepen size={30} />
+              <GoRepo size={30} />
             </div>
             <div>
               <p className="m-0 p-0">Public Repos</p>
-              <p className="m-0 p-0 fw-bold fs-2">
+              <p className="m-0 p-0 fw-bold fs-3">
                 {getGithubUser?.public_repos}
               </p>
             </div>
@@ -169,11 +200,11 @@ export default function UserProfile() {
 
           <div className="col d-flex align-items-center">
             <div className="icon-square text-bg-dark d-inline-flex align-items-center justify-content-center p-2 rounded me-3">
-              <FaStore size={30} />
+              <GoCodeSquare size={30} />
             </div>
             <div>
               <p className="m-0 p-0">Public Gists</p>
-              <p className="m-0 p-0 fw-bold fs-2">
+              <p className="m-0 p-0 fw-bold fs-3">
                 {getGithubUser?.public_gists}
               </p>
             </div>
@@ -181,71 +212,82 @@ export default function UserProfile() {
         </article>
       </section>
 
-      <section className="my-3">
-        <div className="card card-cover shadow p-3">
-          <section className="row">
-            {getGithubUser?.type !== 'Organization' ? (
-              <article className="col-lg-7">
-                <h4 className="text-center my-3 fw-bold">GitHub Statistics</h4>
-                <figure>
-                  <img
-                    src={`https://github-readme-activity-graph.vercel.app/graph?username=${getGithubUser?.login}&bg_color=ffffff&color=000111&line=5194f0&point=000000&area=true&hide_border=true&area_color=D5E5FA`}
-                    alt="Github Activity Graph"
-                    width="100%"
-                    className="img-fluid"
-                  />
-                </figure>
-                <figure>
-                  <img
-                    src={`https://github-profile-trophy.vercel.app/?username=${getGithubUser?.login}&row=1&column=7&margin-w=5&no-frame=false`}
-                    alt="Github Trophies"
-                    width="100%"
-                    className="img-fluid"
-                  />
-                </figure>
+      {getGithubUser.type && getGithubUser?.type === 'User' && (
+        <section className="card card-shadow p-3 my-4">
+          <h2 className="h3 text-center fw-bold mb-4">
+            @{getGithubUser?.login}&#39;s Statistics
+          </h2>
 
-                <figure>
-                  <img
-                    src={`https://github-readme-stats.vercel.app/api/top-langs/?username=${getGithubUser.login}&count_private=true&include_all_commits=true&layout=compact`}
-                    alt="Github Top Languages"
-                    className="img-fluid"
-                    width="100%"
-                  />
-                </figure>
+          <div className="row">
+            <div className="col-12">
+              <figure>
+                <img
+                  src={`https://github-profile-trophy.vercel.app/?username=${getGithubUser?.login}&row=1&margin-w=5&no-frame=false`}
+                  alt="Github Trophies"
+                  width="100%"
+                  className="img-fluid"
+                />
+              </figure>
+            </div>
+            <div className="col-12 col-md-6">
+              <figure>
+                <img
+                  src={`https://github-readme-activity-graph.vercel.app/graph?username=${getGithubUser?.login}&bg_color=ffffff&color=000111&line=5194F0&point=000000&area=true&area_color=5194F0&hide_border=true`}
+                  alt="Github Activity Graph"
+                  width="100%"
+                  className="img-fluid"
+                />
+              </figure>
+            </div>
+            <div className="col-12 col-md-6">
+              <figure>
+                <img
+                  src={`https://github-readme-streak-stats.herokuapp.com/?user=${getGithubUser.login}&ring=5194F0&currStreakLabel=5194F0`}
+                  alt="Github Streaks"
+                  width="100%"
+                  className="img-fluid"
+                />
+              </figure>
+            </div>
+            <div className="col-12 col-md-6">
+              <figure>
+                <img
+                  src={`https://github-contributor-stats.vercel.app/api?username=${getGithubUser.login}&limit=15&combine_all_yearly_contributions=true`}
+                  alt="Github Contribution Statistics"
+                  width="100%"
+                  className="img-fluid"
+                />
+              </figure>
+            </div>
 
-                <figure>
-                  <img
-                    src={`https://github-readme-stats.vercel.app/api?username=${getGithubUser.login}&show_icons=true&count_private=true&include_all_commits=true&show=reviews,discussions_started,discussions_answered,prs_merged,prs_merged_percentage`}
-                    alt="Github Stats"
-                    width="100%"
-                    className="img-fluid"
-                  />
-                </figure>
+            <div className="col-12 col-md-6">
+              <figure>
+                <img
+                  src={`https://github-readme-stats.vercel.app/api?username=${getGithubUser.login}&show_icons=true&count_private=true&include_all_commits=true&show=reviews,prs_merged,prs_merged_percentage`}
+                  alt="Github Overall Statistics"
+                  width="100%"
+                  className="img-fluid"
+                />
+              </figure>
+              <figure>
+                <img
+                  src={`https://github-readme-stats.vercel.app/api/top-langs/?username=${getGithubUser.login}&count_private=true&include_all_commits=true&layout=compact`}
+                  alt="Github Top Languages"
+                  className="img-fluid"
+                  width="100%"
+                />
+              </figure>
+            </div>
+          </div>
+        </section>
+      )}
 
-                <figure>
-                  <img
-                    src={`https://github-readme-streak-stats.herokuapp.com/?user=${getGithubUser.login}`}
-                    alt="Github Streaks"
-                    width="100%"
-                    className="img-fluid"
-                  />
-                </figure>
-              </article>
-            ) : null}
+      <section className="card card-shadow p-3 my-4">
+        <h2 className="h3 text-center fw-bold mb-4">
+          @{getGithubUser?.login}&#39;s Latest Repositories
+        </h2>
 
-            <article
-              className={
-                getGithubUser?.type !== 'Organization'
-                  ? 'col-lg-5'
-                  : 'col-lg-12'
-              }
-            >
-              <h4 className="text-center my-3 fw-bold">Latest Repositories</h4>
-
-              <Repositories getGithubUser={getGithubUser} username={username} />
-            </article>
-          </section>
-        </div>
+        <Repositories getGithubUser={getGithubUser} username={username} />
       </section>
     </React.Fragment>
   )
